requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0
requests==2.31.0
